[REFACTOR]: covered syscall errors
