--- libft_cpp/README.md
@@ -1,2 +1,0 @@
-# libft ver.c++
-
