--- libft_cpp/ft_bzero.cpp
@@ -1,10 +1,0 @@
-#include "libft.hpp"
-
-void	ft_bzero(void *s, size_t n)
-{
-	unsigned char *temp;
-
-	temp = (unsigned char *)s;
-	while (n--)
-		*(temp++) = 0;
-}
