--- libft_cpp/ft_fd_set.cpp
@@ -1,7 +1,0 @@
-#include "libft.hpp"
-
-void FT_FD_SET(int fd, fd_set *p)
-{
-	int mask = 1 << (fd % (sizeof(__int32_t) * 8));
-    p->fds_bits[fd / (sizeof(__int32_t) * 8)] |= mask;
-}
