--- libft_cpp/ft_get_time.cpp
@@ -1,9 +1,0 @@
-#include "libft.hpp"
-
-unsigned long		ft_get_time(void)
-{
-	struct timeval	tv;
-
-	gettimeofday(&tv, NULL);
-	return (tv.tv_sec * 1000 + tv.tv_usec / 1000);
-}
