--- libft_cpp/ft_htons.cpp
@@ -1,11 +1,0 @@
-#include "libft.hpp"
-
-uint16_t ft_htons(uint16_t val)
-{
-	uint16_t temp = 0x1234;
-	if (*(unsigned char *)&temp == 0x12)
-		return (val);
-
-	return ((val & 0x00FFU) << 8
-		| (val & 0xFF00U) >> 8);
-}
