--- libft_cpp/ft_isalnum.cpp
@@ -1,6 +1,0 @@
-#include "libft.hpp"
-
-int	ft_isalnum(int c)
-{
-	return (ft_isalpha(c) || ft_isdigit(c));
-}
