--- libft_cpp/ft_isalpha.cpp
@@ -1,6 +1,0 @@
-#include "libft.hpp"
-
-int	ft_isalpha(int c)
-{
-	return (('A' <= c && c <= 'Z') || ('a' <= c && c <= 'z'));
-}
