--- libft_cpp/ft_isascii.cpp
@@ -1,6 +1,0 @@
-#include "libft.hpp"
-
-int	ft_isascii(int c)
-{
-	return (0 <= c && c <= 127);
-}
