--- libft_cpp/ft_isdigit.cpp
@@ -1,6 +1,0 @@
-#include "libft.hpp"
-
-int	ft_isdigit(int c)
-{
-	return ('0' <= c && c <= '9');
-}
