--- libft_cpp/ft_isprint.cpp
@@ -1,6 +1,0 @@
-#include "libft.hpp"
-
-int	ft_isprint(int c)
-{
-	return (32 <= c && c <= 126);
-}
