--- libft_cpp/ft_nbr_length.cpp
@@ -1,16 +1,0 @@
-#include "libft.hpp"
-
-int		ft_nbr_length(int n)
-{
-	int i;
-
-	i = 0;
-	if (n <= 0)
-		i++;
-	while (n != 0)
-	{
-		n /= 10;
-		i++;
-	}
-	return (i);
-}
