--- libft_cpp/ft_putchar_fd.cpp
@@ -1,8 +1,0 @@
-#include "libft.hpp"
-
-void	ft_putchar_fd(char c, int fd)
-{
-	if (fd < 0)
-		return ;
-	write(fd, &c, 1);
-}
