--- libft_cpp/ft_putnbr_fd.cpp
@@ -1,21 +1,0 @@
-#include "libft.hpp"
-
-void	ft_putnbr_fd(int n, int fd)
-{
-	if (fd < 0)
-		return ;
-	if (n < 0)
-	{
-		write(fd, "-", 1);
-		if (n == -2147483648)
-		{
-			write(fd, "2147483648", 10);
-			return ;
-		}
-		ft_putnbr_fd(-n, fd);
-		return ;
-	}
-	if (n > 9)
-		ft_putnbr_fd(n / 10, fd);
-	ft_putchar_fd(n % 10 + 48, fd);
-}
