--- libft_cpp/ft_tolower.cpp
@@ -1,8 +1,0 @@
-#include "libft.hpp"
-
-int	ft_tolower(int c)
-{
-	if ('A' <= c && c <= 'Z')
-		c += 32;
-	return (c);
-}
