--- libft_cpp/ft_toupper.cpp
@@ -1,8 +1,0 @@
-#include "libft.hpp"
-
-int	ft_toupper(int c)
-{
-	if ('a' <= c && c <= 'z')
-		c -= 32;
-	return (c);
-}
